Address PR review: deduplicate helper, super call, safer index
- Move `makeSwiftUIHostingController` to AccessibilitySnapshotPreviews as
  public, used by both SnapshotTesting+SwiftUI and FBSnapshotTestCase+SwiftUI
  instead of being copy-pasted in each.
- Call `super.layoutSubviews()` in SwiftUIAccessibilitySnapshotContainerView.
- Replace `cells[0].count` with `cells.first?.count ?? 0` in
  DataTableContainerDemo so the description string degrades gracefully if
  the demo data is ever edited to be empty.
